fix(#2229): stop truncating 64-bit connect and idle deltas to int
CLinearTimeDelta::ReturnSeconds() is 64-bit, but the reporting path narrowed it
to int, so a delta past 2^31 wrapped into a plausible smaller number instead of
an obviously wrong one. That is how 2.13's #2227 stayed hidden: a live
descriptor whose conn(<port>) is 6,220,867,076 s -- 197 years -- shows in WHO as
22290d 11h, exactly that value less 2^32.

2.14 already widened the softcode entry points (mux_atoi64 in fun_singletime and
friends) but narrowed the value straight back at time_format_2, so singletime()
was wrong anyway.

The formatters cross the driver/engine boundary, so this is an ABI change:
TimeFormat1 and TimeFormat2 on mux_IGameEngine, and FetchIdle and FetchConnect
on mux_IConnectionManager, all widen. IID_IGameEngine and
IID_IConnectionManager are bumped accordingly, and both implementations of the
connection manager -- netmux's CConnectionManager and muxscript's
CScriptConnectionManager -- move together.

Also widened GeneralTimeConversion and its int iValue accumulator. With a
64-bit input the term-by-term remainder subtraction overflowed and every term
printed the whole span rather than what was left of it; exptime() answered
"197y 1634M 7104w 49713d 1193069h 71582803m 1s". The buffer-size comments there
already reason in terms of 2^63, so 64-bit was the original intent.

tf1_width_table used INT_MAX as the end-of-cases sentinel for widths 10 and 11.
Widening the parameter alone would let a large value walk past the final case
into case 3, whose format spec is empty at those widths. The real range maxima
from the comment block replace it, and a value too large for the available
columns now yields "???" -- which the function already emits for a bad
maxWidth. Rendering an impossible duration as a believable one is the failure
being fixed; a marker is the point.

Verified against a running server, with #2227's corruption injected on the
driver side so the oversized value has to survive the vtable crossing:

  singletime(6220596985)  61y  -> 197y
  exptime(6220596985)     cumulative junk -> 197y 1M 2w 12h 20s
  WHO On For              22290d 11h -> 71997d 15h

197y 1M 2w 12h 20s decomposes back to 6,220,596,985 exactly. Ordinary values are
unchanged: singletime(90061) is 1d and a fresh session still reads 0:00.
engine.so loads against the bumped IIDs.

make test: 34 passed, 2 skipped, 0 failed (jit=yes stubslave=yes).

+++ mux/modules/engine/conn_bridge.cpp
@@ -230,16 +230,16 @@ int fetch_width(dbref target)
     return n;
 }
 
-int fetch_idle(dbref target)
+int64_t fetch_idle(dbref target)
 {
-    int n = -1;
+    int64_t n = -1;
     if (g_pConnMgr) g_pConnMgr->FetchIdle(target, &n);
     return n;
 }
 
-int fetch_connect(dbref target)
+int64_t fetch_connect(dbref target)
 {
-    int n = -1;
+    int64_t n = -1;
     if (g_pConnMgr) g_pConnMgr->FetchConnect(target, &n);
     return n;
 }
